build: check if dpdk submodule is initialized

+++ build.sh
@@ -13,7 +13,14 @@ then
 fi
 make -j 8 'CFLAGS=-DLUAJIT_NUMMODE=2 -DLUAJIT_ENABLE_LUA52COMPAT'
 make install DESTDIR=$(pwd)
+
 cd ../dpdk
+if [[ ! -e Makefile ]]
+then
+	echo "ERROR: DPDK submodule not initialized"
+	echo "Please run git submodule update --init"
+	exit 1
+fi
 make -j 8 install T=x86_64-native-linuxapp-gcc
 ../../bind-interfaces.sh
 cd ../../build
